ai_engine/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.10.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0

# AI SDKs
anthropic>=0.40.0
google-generativeai>=0.3.2

# Database & Cache
redis>=5.0.1

# HTTP & Tools
httpx>=0.26.0
requests>=2.31.0
aiohttp>=3.9.1

# HTML Parsing
beautifulsoup4>=4.12.3

# Monitoring & Logging
structlog>=24.1.0
sentry-sdk[fastapi]>=1.40.0
slowapi>=0.1.9

# Browser Automation (Browserbase)
browserbase>=1.0.0
playwright>=1.40.0
